refactor(components): 优化组件装配和查询系统
- 使用 getFirst() 替代 get(0) 提升代码简洁性
- 更新 ComponentStorage.loadComponent 调用以传递正确的 worldDir 参数
- 将标签匹配逻辑重构为私有方法提高代码复用
- 修改几何尺寸比较使用扁平化属性提升性能
- 调整连续放置引擎和套接字处理器的导入路径
- 修复组件体素化中的深度计算错误
- 优化墙开口检测算法减少不必要的变量声明
- 更新记录类访问方式使用括号方法调用
- 调整构造函数参数以明确指定切线方向为null
- 修复锚点解析中的坐标转换精度问题
- 重构套接字匹配评分逻辑增强匹配准确性
- 优化路径和轮廓工具数据访问接口一致性
- 移除未使用的方向枚举值和冗余变量声明

+++ src/main/java/com/formacraft/common/component/socket/place/SocketAnchorResolver.java
@@ -58,9 +58,9 @@ public final class SocketAnchorResolver {
 
         // 4️⃣ Patch 原点 = socketCenter - localOffset
         BlockPos origin = new BlockPos(
-                Math.floor(socketCenter.x) - localOffset.getX(),
-                Math.floor(socketCenter.y) - localOffset.getY(),
-                Math.floor(socketCenter.z) - localOffset.getZ()
+                (int) Math.floor(socketCenter.x) - localOffset.getX(),
+                (int) Math.floor(socketCenter.y) - localOffset.getY(),
+                (int) Math.floor(socketCenter.z) - localOffset.getZ()
         );
 
         // 5️⃣ 镜像（v1：默认不开，交给 SymmetryTool）
